Fix parsing bugs and move keyword checkers to Command classes

+++ src/main/java/seedu/vms/logic/commands/keyword/DeleteCommand.java
@@ -1,6 +1,7 @@
 package seedu.vms.logic.commands.keyword;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.vms.commons.core.Messages.MESSAGE_KEYWORD_DOES_NOT_EXIST;
 
 import java.util.Objects;
 
@@ -9,6 +10,7 @@ import seedu.vms.logic.commands.Command;
 import seedu.vms.logic.commands.exceptions.CommandException;
 import seedu.vms.model.Model;
 import seedu.vms.model.keyword.Keyword;
+import seedu.vms.model.keyword.KeywordManager;
 
 /**
  * Deletes a keyword identified using it's displayed index from the keyword manager.
@@ -33,6 +35,9 @@ public class DeleteCommand extends Command {
     @Override
     public CommandMessage execute(Model model) throws CommandException {
         requireNonNull(model);
+        if (!KeywordManager.existingMappingExists(targetKeyword)) {
+            throw new CommandException(String.format(MESSAGE_KEYWORD_DOES_NOT_EXIST, targetKeyword));
+        }
         Keyword keywordToDelete = model.deleteKeyword(targetKeyword);
         return new CommandMessage(String.format(MESSAGE_DELETE_KEYWORD_SUCCESS, keywordToDelete));
     }
